backend/requirements.txt
# Existing dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
qrcode[pil]==7.4.2
opencv-python==4.8.1.78
pyzbar==0.1.9
numpy==1.26.2
python-multipart==0.0.6
Pillow==10.1.0

# 🆕 NEW: ML/DL dependencies
ultralytics==8.1.0          # YOLOv8
torch==2.1.0                # PyTorch
torchvision==0.16.0         # Image transforms
scikit-learn==1.3.2         # Random Forest
pandas==2.1.3               # Data manipulation
matplotlib==3.8.2           # Visualization
seaborn==0.13.0             # Statistical plots
jupyter==1.0.0              # Notebooks
joblib==1.3.2               # Model serialization
tqdm==4.66.1                # Progress bars